RT-N18U: GPL source release 3.0.0.4.378.4850

+++ release/src-rt/platform.mak
@@ -1,5 +1,40 @@
 export LINUXDIR := $(SRCBASE)/linux/linux-2.6
 
+ifeq ($(RTCONFIG_BCMARM),y)
+ifeq ($(EXTRACFLAGS),)
+export EXTRACFLAGS := -DBCMWPA2 -DBCMARM -fno-delete-null-pointer-checks -marm
+endif
+
+export KERNEL_BINARY=$(LINUXDIR)/arch/arm/boot/zImage
+export PLATFORM := arm-uclibc
+export CROSS_COMPILE := arm-brcm-linux-uclibcgnueabi-
+export CROSS_COMPILER := $(CROSS_COMPILE)
+export CONFIGURE := ./configure --host=arm-linux --build=$(BUILD)
+export HOSTCONFIG := linux-armv4
+export BCMEX := _arm
+export EXTRA_FLAG := -lgcc_s
+export ARCH := arm
+export HOST := arm-linux
+export TOOLS := $(SRCBASE)/toolchains/hndtools-arm-linux-2.6.36-uclibc-4.5.3
+export RTVER := 0.9.32.1
+export BCMSUB := brcmarm
+else
+ifeq ($(EXTRACFLAGS),)
+export EXTRACFLAGS := -DBCMWPA2 -fno-delete-null-pointer-checks -mips32 -mtune=mips32
+endif
+
+export KERNEL_BINARY=$(LINUXDIR)/arch/mips/brcm-boards/bcm947xx/compressed/zImage
+export PLATFORM := mipsel-uclibc
+export CROSS_COMPILE := mipsel-uclibc-
+export CROSS_COMPILER := $(CROSS_COMPILE)
+export READELF := mipsel-linux-readelf
+export CONFIGURE := ./configure --host=mipsel-linux --build=$(BUILD)
+export HOSTCONFIG := linux-mipsel
+export ARCH := mips
+export HOST := mipsel-linux
+export TOOLS := $(SRCBASE)/../../tools/brcm/hndtools-mipsel-linux
+export RTVER := 0.9.30.1
+endif
 EXTRA_CFLAGS := -DLINUX26 -DCONFIG_BCMWL5 -DDEBUG_NOISY -DDEBUG_RCTEST -pipe -DTTEST
 
 export CONFIG_LINUX26=y
@@ -45,17 +80,23 @@ define platformKernelConfig
 		mkdir -p $(SRCBASE)/router/ctf_arm/linux;\
 		cp -f $(SRCBASE)/router/ctf_arm/bcm6x/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
 		if [ "$(BCM7)" = "y" ]; then \
-			cp -f $(SRCBASE)/../../dhd/src/shared/rtecdc_43602a0.h.in $(SRCBASE)/../../dhd/src/shared/rtecdc_43602a0.h;\
-			cp -f $(SRCBASE)/../../dhd/src/shared/rtecdc_43602a1.h.in $(SRCBASE)/../../dhd/src/shared/rtecdc_43602a1.h;\
 			if [ "$(ARMCPUSMP)" = "up" ]; then \
 				cp -f $(SRCBASE)/router/ctf_arm/bcm7_up/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
 			else \
-				if [ "$(DPSTA)" = "y" ]; then \
-					cp -f $(SRCBASE)/router/ctf_arm/bcm7_dpsta/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
-					cp -f $(SRCBASE)/router/dpsta/bcm7_3200/dpsta.o $(SRCBASE)/router/dpsta/linux;\
-				else \
-					cp -f $(SRCBASE)/router/ctf_arm/bcm7/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
-				fi; \
+				cp -f $(SRCBASE)/router/ctf_arm/bcm7/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
+				cp -f $(SRCBASE)/router/dpsta/bcm7_3200/dpsta.o $(SRCBASE)/router/dpsta/linux;\
+			fi; \
+		elif [ "$(BCM794)" = "y" ]; then \
+			if [ "$(ARMCPUSMP)" = "up" ]; then \
+				cp -f $(SRCBASE)/router/ctf_arm/bcm7_up/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
+			else \
+				cp -f $(SRCBASE)/router/ctf_arm/bcm7/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
+			fi; \
+		elif [ "$(BCM10)" = "y" ]; then \
+			if [ "$(ARMCPUSMP)" = "up" ]; then \
+				cp -f $(SRCBASE)/router/ctf_arm/bcm7_up/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
+			else \
+				cp -f $(SRCBASE)/router/ctf_arm/bcm7/ctf.* $(SRCBASE)/router/ctf_arm/linux/;\
 			fi; \
 		else \
 			if [ "$(ARMCPUSMP)" = "up" ]; then \
@@ -185,22 +226,42 @@ define platformKernelConfig
 		echo "# CONFIG_CRYPTO_CBC is not set" >> $(1); \
 	fi; \
 	if [ "$(ARM)" = "y" ]; then \
-		if [ -d $(SRCBASE)/router/wl_arm/prebuilt ]; then \
-			mkdir $(SRCBASE)/wl/linux ; \
-			cp $(SRCBASE)/router/wl_arm/prebuilt/* $(SRCBASE)/wl/linux ; \
-		elif [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME) ]; then \
-			if [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/linux ]; then \
-				cp -rf $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/linux $(SRCBASE)/wl/. ; \
+		if [ "$(BCM7)" = "y" ]; then \
+			if [ -d $(SRCBASE)/../../43602/src/wl/sysdeps/$(BUILD_NAME) ]; then \
+				if [ -d $(SRCBASE)/../../43602/src/wl/sysdeps/$(BUILD_NAME)/clm ]; then \
+					cp -f $(SRCBASE)/../../43602/src/wl/sysdeps/$(BUILD_NAME)/clm/src/wlc_clm_data.c $(SRCBASE)/../../43602/src/wl/clm/src/. ; \
+					cp -f $(SRCBASE)/../../43602/src/wl/sysdeps/$(BUILD_NAME)/clm/src/wlc_clm_data_inc.c $(SRCBASE)/../../43602/src/wl/clm/src/. ; \
+				fi; \
+			else \
+				if [ -d $(SRCBASE)/../../43602/src/wl/sysdeps/default/clm ]; then \
+					cp -f $(SRCBASE)/../../43602/src/wl/sysdeps/default/clm/src/wlc_clm_data.c $(SRCBASE)/../../43602/src/wl/clm/src/. ; \
+					cp -f $(SRCBASE)/../../43602/src/wl/sysdeps/default/clm/src/wlc_clm_data_inc.c $(SRCBASE)/../../43602/src/wl/clm/src/. ; \
+				fi; \
 			fi; \
-			if [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/clm ]; then \
-				cp -f $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/clm/src/wlc_clm_data.c $(SRCBASE)/wl/clm/src/. ; \
+			if [ -d $(SRCBASE)/router/wl_arm_7/prebuilt ]; then \
+				mkdir $(SRCBASE)/wl/linux ; \
+				cp $(SRCBASE)/router/wl_arm_7/prebuilt/wl_apsta.o $(SRCBASE)/wl/linux ; \
+				mkdir -p $(SRCBASE)/../../dhd/src/dhd/linux ; \
+				cp $(SRCBASE)/router/wl_arm_7/prebuilt/dhd.o $(SRCBASE)/../../dhd/src/dhd/linux ; \
 			fi; \
 		else \
-			if [ -d $(SRCBASE)/wl/sysdeps/default/linux ]; then \
-				cp -rf $(SRCBASE)/wl/sysdeps/default/linux $(SRCBASE)/wl/. ; \
-			fi; \
-			if [ -d $(SRCBASE)/wl/sysdeps/default/clm ]; then \
-				cp -f $(SRCBASE)/wl/sysdeps/default/clm/src/wlc_clm_data.c $(SRCBASE)/wl/clm/src/. ; \
+			if [ -d $(SRCBASE)/router/wl_arm/prebuilt ]; then \
+				mkdir $(SRCBASE)/wl/linux ; \
+				cp $(SRCBASE)/router/wl_arm/prebuilt/wl_apsta.o $(SRCBASE)/wl/linux ; \
+			elif [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME) ]; then \
+				if [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/linux ]; then \
+					cp -rf $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/linux $(SRCBASE)/wl/. ; \
+				fi; \
+				if [ -d $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/clm ]; then \
+					cp -f $(SRCBASE)/wl/sysdeps/$(BUILD_NAME)/clm/src/wlc_clm_data.c $(SRCBASE)/wl/clm/src/. ; \
+				fi; \
+			else \
+				if [ -d $(SRCBASE)/wl/sysdeps/default/linux ]; then \
+					cp -rf $(SRCBASE)/wl/sysdeps/default/linux $(SRCBASE)/wl/. ; \
+				fi; \
+				if [ -d $(SRCBASE)/wl/sysdeps/default/clm ]; then \
+					cp -f $(SRCBASE)/wl/sysdeps/default/clm/src/wlc_clm_data.c $(SRCBASE)/wl/clm/src/. ; \
+				fi; \
 			fi; \
 		fi; \
 	else \
